devtools/client/shared/components/reps/test/mochitest/test_reps_infinity.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test Infinity rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - Infinity</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
"use strict";

window.onload = Task.async(function* () {
  const {
    REPS,
    getRep,
  } = browserRequire("devtools/client/shared/components/reps/reps");
  let { InfinityRep } = REPS;

  try {
    yield testInfinity();
    yield testNegativeInfinity();
  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }

  function testInfinity() {
    const stub = getGripStub("testInfinity");
    is(getRep(stub), InfinityRep.rep, "Rep correctly selects Infinity Rep");

    const renderedComponent = renderComponent(InfinityRep.rep, { object: stub });
    is(renderedComponent.textContent, "Infinity",
      "Infinity rep has expected text content for Infinity");
  }

  function testNegativeInfinity() {
    const stub = getGripStub("testNegativeInfinity");
    is(getRep(stub), InfinityRep.rep, "Rep correctly selects Infinity Rep");

    const renderedComponent = renderComponent(InfinityRep.rep, { object: stub });
    is(renderedComponent.textContent, "-Infinity",
      "Infinity rep has expected text content for negative Infinity");
  }

  function getGripStub(name) {
    switch (name) {
      case "testInfinity":
        return {
          type: "Infinity"
        };
      case "testNegativeInfinity":
        return {
          type: "-Infinity"
        };
    }
    return null;
  }
});
</script>
</pre>
</body>
</html>
